Add two more parameters too RIL_REQUEST_SETUP_DATA_CALL, DO NOT MERGE.
Add DNS and gateway as parameters to RIL_REQUEST_SETUP_DATA_CALL so that
system properties aren't used for passing these to the ril. System
properties that are read right after a write doesn't always see the
result of the write. Another words writing system properties are not
not sequentially consistent.

Bug: 3364487
Change-Id: I408e56cd8cf943dfb5d71a21a274fcfd14571b7b