Use local copy of ril.h for reference-ril and mock-ril.
This allows for greater stability while we're changing the ril
interface, as defined by ril.h. An example is that the emulator
will continue to run using an older ril while we upgrade to
newer rils.

After the new ril has stabilized and these modules updated the
changes should be reverted and we should use
include/telephony/ril.h

Change-Id: If5011ccbe64bcb65f98b02060165f899b6d4b5f9