Fix responseSimStatus use of responselen.
Previously this code was checking that the responselen was a
multiple of the length of a pointer. I'd changed it to assume
it was supposed to be a pointer. But actually the reference-ril
sets it to the length of the object. So fixing this to assume
its the length of the object.

Change-Id: I90e9fdee89151359041eccab3cfd2d442c7fbcbf