reference-ril: Stop using pthread_cond_timedwait_relative_np()
pthread_cond_timedwait_relative_np() is deprecated, and is only
available in the 32-bit ABI (see bionic/libc/include/pthread.h). It is
still used by the 32-bit version of reference-ril. Thanks to a recent
bugfix for 64-bit [1], it is now safe to replace the deprecated
function with the standard pthread_cond_timedwait(). Doing so also
simplifies the code, as there is now a unified logic for 32 and 64-bit
targets.

[1] https://android-review.googlesource.com/210301

Change-Id: I0895bbf65ddb71965018d6945dfb8baa786d5d89
Signed-off-by: Yu Ning <yu.ning@intel.com>