reference-ril: Stop using pthread_cond_timedwait_relative_np()
pthread_cond_timedwait_relative_np() is deprecated, and is only
available in the 32-bit ABI (see bionic/libc/include/pthread.h). It is
still used by the 32-bit version of reference-ril. Thanks to a recent
bugfix for 64-bit [1], it is now safe to replace the deprecated
function with the standard pthread_cond_timedwait(). Doing so also
simplifies the code, as there is now a unified logic for 32 and 64-bit
targets.

[1] https://android-review.googlesource.com/210301

Signed-off-by: Yu Ning <yu.ning@intel.com>

(cherry picked from commit 3480b1832d3581c708868117c0eb21ba291df165)

Change-Id: I1a15ee0fdfea38b6ee9c9fd49845d10d45f485b1