Use init to set capabilities (re-land)
(This is a re-land now that
https://android-review.googlesource.com/#/c/314804/ has fixed the
underlying bug. The problem was that the 3.10 kernel on angler
didn't support some newer capabilities. This CL fixes a bug reported by
Sony.)

With the minijail code the groups set in the rild.rc are lost when
switching user. The reason for starting as root was to be able to set
capabilities, since that code was written init has gain the ability
to set capabilities at service declaration so use that instead so we
dont even have to start as root.

Before change:

Name:	rild
...
Uid:	1001	1001	1001	1001
Gid:	1001	1001	1001	1001
...
Groups:
...
CapInh: 0000001000003000
CapPrm: 0000001000003000
CapEff: 0000001000003000
CapBnd: 0000001000003000
CapAmb: 0000000000000000
...

After change:

Name:   rild
...
Uid:    1001    1001    1001    1001
Gid:    1001    1001    1001    1001
...
Groups: 1005 1007 2001 3003 3009 3010 9998
...
CapInh: 0000001000003000
CapPrm: 0000001000003000
CapEff: 0000001000003000
CapBnd: 0000001000003000
CapAmb: 0000001000003000
...

Bug: 33443755
Test: Build/boot/check rild is running on angler and marlin.
Change-Id: I138f6203e55ea2c304e506da3949b5a932b83401