Allow newer message format indicating signal strength
This code expected v6 of the cellular signal strength
message. The framework allows up to v10, which takes
more memory. In fact, the framework requires that it
be given a buffer that is the right size for a v10
message.

This code now uses a buffer that is sized for v10.
It accepts either a v6 or a v10 message.

Bug: b.android.com/316334

Test: Manually verified correct behavior with Emulator.

Change-Id: Iaf8dda0e441393b6fde899fa30bcc90a4413f0ef