Use system property gateway if not using WiFi
When running with WiFi the IPv4 gateway for the radio interface is known
and can be hardcoded, even if TAP is enabled. This is because the
gateway is the internal address in the router namespace. Without WiFi
the radio interface is eth0, the public interface. The gateway for this
interface can be dynamic if running with TAP networking so in that case
it can't use a hardcoded gateway the way it used to. The DHCP client
will set the net.eth0.gw system property to the gateway of the host
network so use that property to give the correct gateway to the radio
interface.
Bug: 72886046
Test: manual, build and ran the emulator (steps from CTS tests)
Change-Id: I110a1216cccc846a48befe4c8cb2e602ed63752f
(cherry picked from commit cf111309af68ab095371ef495d2c78e1faf885a6)
(cherry picked from commit 0de94de9dc67db550f50bf2ffe3360b7ce02907d)